게시물 CREATE, UPDATE, DELETE 작업 시 로그아웃 상태 예외처리

+++ server/controllers/postController.js
@@ -20,33 +20,38 @@ export const getUploadPost = (req, res) => {
 }
 
 export const uploadPost = async (req, res) => {
-  const {
-    user: { _id },
-  } = req.session;
-  if(!req.body.제목 || !req.body.내용){
-    res.status(400).send('제목 혹은 내용이 입력되지 않았습니다.');
-    return;
+  const isLoggedIn = req.session.loggedIn;
+  if(isLoggedIn){
+    const {
+      user: { _id },
+    } = req.session;
+    if(!req.body.제목 || !req.body.내용){
+      res.status(400).send('제목 혹은 내용이 입력되지 않았습니다.');
+      return;
+    }
+    const {제목, 내용} = req.body;
+  
+    try{
+      const newPost = await Post.create({
+        제목,
+        내용,
+        createdAt:Date.now(),
+        owner: _id,
+        meta:{
+          views:0,
+        }
+      })
+      const user = await User.findById(_id);
+      user.posts.push(newPost._id);
+      user.save();
+      return res.send("uploaded completely");
+    } catch (error) {
+      return res.status(400).send("에러");
+    }
+  } else {
+    return res.status(401).send("로그인이 필요합니다.")
   }
-  const {제목, 내용} = req.body;
-
-  try{
-    const newPost = await Post.create({
-      제목,
-      내용,
-      createdAt:Date.now(),
-      owner: _id,
-      meta:{
-        views:0,
-      }
-    })
-    const user = await User.findById(_id);
-    user.posts.push(newPost._id);
-    user.save();
-    return res.send("uploaded completely");
-  } catch (error) {
-    return res.status(400).send("에러");
-  } finally {
-  }
+  
   
   
 };
@@ -60,44 +65,54 @@ export const getEditPost = (req, res) => {
 }
 
 export const editPost = async (req, res) => {
-  const {
-    user: { _id },
-  } = req.session;
-  const {id} = req.params;
-  const {제목,내용} = req.body;
-  const post = await Post.findById(id);
-
-  if (!post) {
-    return res.status(404).send("Not Found");
+  const isLoggedIn = req.session.loggedIn;
+  if(isLoggedIn){
+    const {
+      user: { _id },
+    } = req.session;
+    const {id} = req.params;
+    const {제목,내용} = req.body;
+    const post = await Post.findById(id);
+  
+    if (!post) {
+      return res.status(404).send("Not Found");
+    }
+  
+    if (String(post.owner) !== String(_id)) {
+      return res.status(403).send("해당 게시물의 작성자가 아닙니다.");
+    }
+  
+    await Post.findByIdAndUpdate(id, {
+      제목,
+      내용
+    });
+    return res.json(post);
+  } else {
+    return res.status(401).send("로그인이 필요합니다.")
   }
-
-  if (String(post.owner) !== String(_id)) {
-    return res.status(403).send("해당 게시물의 작성자가 아닙니다.");
-  }
-
-  await Post.findByIdAndUpdate(id, {
-    제목,
-    내용
-  });
-
-  return res.json(post);
 };
 
 export const deletePost = async (req, res) => {
-  const {
-    user: { _id },
-  } = req.session;
-  const { id } = req.params;
-
-  const post = await Post.findById(id);
-
-  if (!post) {
-    return res.status(404).send("Post not found.");
+  const loggedIn = req.session.loggedIn;
+  if(loggedIn){
+    const {
+      user: { _id },
+    } = req.session;
+    const { id } = req.params;
+  
+    const post = await Post.findById(id);
+  
+    if (!post) {
+      return res.status(404).send("Post not found.");
+    }
+    if (String(post.owner) !== String(_id)) {
+      return res.status(403).send("해당 게시물의 작성자가 아닙니다.");
+    }
+  
+    await Post.findByIdAndDelete(id);
+    return res.send("post's deleted completely");
+  } else {
+    return res.status(401).send("로그인이 필요합니다.")
   }
-  if (String(post.owner) !== String(_id)) {
-    return res.status(403).send("해당 게시물의 작성자가 아닙니다.");
-  }
-
-  await Post.findByIdAndDelete(id);
-  return res.send("post's deleted completely");
+  
 };
